Move IdStatus to frontend
