+++ qsu/src/test/scala/quasar/qsu/ResolveOwnIdentitiesSpec.scala
@@ -19,12 +19,13 @@ package quasar.qsu
 import slamdata.Predef._
 
 import quasar.{Qspec, TreeMatchers}
+import quasar.IdStatus.{ExcludeId, IdOnly, IncludeId}
 import quasar.contrib.pathy.AFile
 import quasar.ejson.EJson
 import quasar.ejson.implicits._
 import quasar.contrib.iota.{copkEqual, copkTraverse}
 import quasar.fp.ski.κ
-import quasar.qscript.{construction, ExcludeId, Hole, IdOnly, IncludeId, OnUndefined, SrcHole}
+import quasar.qscript.{construction, Hole, OnUndefined, SrcHole}
 
 import matryoshka._
 import matryoshka.data.Fix
